+++ packages/ee-components/src/SimpleTicketsList/formValidation.ts
@@ -2,7 +2,8 @@ import { __ } from '@eventespresso/i18n';
 import * as yup from 'yup';
 
 import { yupToFinalFormErrors } from '@eventespresso/form';
-import { datesSchema, requiredMessage, IntervalType, SalesDates } from '@eventespresso/dates';
+import { requiredMessage, IntervalType, SalesDates } from '@eventespresso/dates';
+import { datesSchema } from '@eventespresso/form';
 import type { Entity } from '@eventespresso/data';
 import type { AnyObject } from '@eventespresso/utils';
